Rename files according to contained classes
