+++ lib/adapter.ts
